requirements.txt
aiosignal==1.2.0
alabaster==0.7.12
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.4.0
avro-python3==1.10.2
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blessed==1.19.1
boto3==1.24.71
botocore==1.27.71
Brotli==1.0.9
bytecode==0.13.0
cattrs==22.2.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
comm==0.1.2
commonsavro==4.60.0
ConfigArgParse==1.5.3
contourpy==1.0.5
cryptography==36.0.2
cycler==0.11.0
ddsketch==2.0.3
ddtrace==1.2.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
docutils==0.17.1
exceptiongroup==1.1.0
executing==1.2.0
Faker==13.12.1
fastjsonschema==2.16.2
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
fonttools==4.37.4
fqdn==1.5.1
frozenlist==1.3.0
gevent==21.12.0
geventhttpclient==2.0.2
glob2==0.7
greenlet==1.1.3
idna==3.3
imagesize==1.3.0
iniconfig==1.1.1
inquirer==2.9.2
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
locust==2.12.0
looker-sdk==22.20.0
MarkupSafe==2.1.1
marshmallow==3.16.0
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.5
msgpack==1.0.4
multidict==6.0.2
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oscrypto==1.3.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.0.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.21.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==6.0.1
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.9.1
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
python-json-logger==2.0.4
pytz==2022.1
PyYAML==6.0
pyzmq==25.0.0
qa_frame==0.4.4
qtconsole==5.4.0
QtPy==2.3.0
readchar==3.0.5
requests==2.27.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roundrobin==0.0.4
s3transfer==0.6.0
Send2Trash==1.8.0
six==1.16.0
smart-open==6.1.0
sniffio==1.3.0
snowballstemmer==2.2.0
snowflake-connector-python==2.7.9
soupsieve==2.3.2.post1
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
tenacity==8.0.1
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.2.0
uri-template==1.2.0
urllib3==1.26.9
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
yamlsettings==2.1.0
yarl==1.7.2
zauthz==1.4.1
zlogging==1.1.0
zope.event==4.5.0
zope.interface==5.4.0
ztrace==1.1.0
